+++ src/main/java/com/box/sdk/BoxResourceIterable.java
@@ -53,19 +53,59 @@ public abstract class BoxResourceIterable<T> implements Iterable<T> {
     private final int limit;
 
     /**
+     * The iterator that gets the next items.
+     */
+    private final IteratorImpl iterator;
+
+    /**
      * Constructor.
      *
-     * @param api
-     *            the API connection to be used by the resource
-     * @param url
-     *            to end-point with paging support
-     * @param limit
-     *            the maximum number of items to return in a page
+     * @param api the API connection to be used by the resource
+     * @param url endpoint with paging support
+     * @param limit the maximum number of items to return in a page
      */
     public BoxResourceIterable(BoxAPIConnection api, URL url, int limit) {
+        this(api, url, limit, null, null);
+    }
+
+    /**
+     * Constructor.
+     *
+     * @param api the API connection to be used by the resource.
+     * @param url to endpoint with paging support.
+     * @param limit the maximum number of items to return in a page.
+     * @param body the body to send to the requested endpoint.
+     */
+    public BoxResourceIterable(BoxAPIConnection api, URL url, int limit, JsonObject body) {
+        this(api, url, limit, body, null);
+    }
+
+    /**
+     * Constructor.
+     *
+     * @param api  the API connection to be used by the resource.
+     * @param url  to endpoint with paging support.
+     * @param limit the maximum number of items to return in a page.
+     * @param marker the marker where the iterator will begin
+     */
+    public BoxResourceIterable(BoxAPIConnection api, URL url, int limit, String marker) {
+        this(api, url, limit, null, marker);
+    }
+
+    /**
+     * Constructor.
+     *
+     * @param api  the API connection to be used by the resource.
+     * @param url  to endpoint with paging support.
+     * @param limit the maximum number of items to return in a page.
+     * @param body the body to send to the requested endpoint.
+     * @param marker the marker where the iterator will begin
+     */
+    public BoxResourceIterable(BoxAPIConnection api, URL url, int limit, JsonObject body, String marker) {
         this.api = api;
         this.url = url;
         this.limit = limit;
+        this.iterator = new IteratorImpl(marker, body);
     }
 
     /**
@@ -85,7 +125,17 @@ public abstract class BoxResourceIterable<T> implements Iterable<T> {
      */
     @Override
     public Iterator<T> iterator() {
-        return new IteratorImpl();
+        return this.iterator;
+    }
+
+    /**
+     * Builds internal read-only iterator over {@link BoxResource}-s.
+     *
+     * @return iterator implementation
+     * @see Iterable#iterator()
+     */
+    public String getNextMarker() {
+        return this.iterator.markerNext;
     }
 
     /**
@@ -110,9 +160,19 @@ public abstract class BoxResourceIterable<T> implements Iterable<T> {
         private int pageCursor;
 
         /**
-         * Constructor.
+         * The body to include in the request.
          */
-        IteratorImpl() {
+        private JsonObject body;
+
+        /**
+         * Constructor.
+         *
+         * @param marker
+         *            the marker at which the iterator will begin
+         */
+        IteratorImpl(String marker, JsonObject body) {
+            this.markerNext = marker;
+            this.body = body;
             this.loadNextPage();
         }
 
@@ -124,7 +184,11 @@ public abstract class BoxResourceIterable<T> implements Iterable<T> {
             QueryStringBuilder builder = new QueryStringBuilder(existingQuery);
             builder.appendParam(PARAMETER_LIMIT, BoxResourceIterable.this.limit);
             if (this.markerNext != null) {
-                builder.appendParam(PARAMETER_MARKER, this.markerNext);
+                if (this.body != null) {
+                    this.body.set("marker", this.markerNext);
+                } else {
+                    builder.appendParam(PARAMETER_MARKER, this.markerNext);
+                }
             }
 
             URL url;
@@ -134,7 +198,15 @@ public abstract class BoxResourceIterable<T> implements Iterable<T> {
                 throw new BoxAPIException("Couldn't append a query string to the provided URL.");
             }
 
-            BoxAPIRequest request = new BoxAPIRequest(BoxResourceIterable.this.api, url, "GET");
+            BoxAPIRequest request = null;
+            if (this.body != null) {
+                request = new BoxAPIRequest(BoxResourceIterable.this.api, url, "POST");
+                request.setBody(this.body.toString());
+                request.addHeader("Content-Type", "application/json");
+            } else {
+                request = new BoxAPIRequest(BoxResourceIterable.this.api, url, "GET");
+            }
+
             BoxJSONResponse response = (BoxJSONResponse) request.send();
             JsonObject pageBody = JsonObject.readFrom(response.getJSON());
 
@@ -184,7 +256,6 @@ public abstract class BoxResourceIterable<T> implements Iterable<T> {
         public void remove() {
             throw new UnsupportedOperationException();
         }
-
     }
 
 }