+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2.42.0 [2019-12-17]
+- Added Metadata Query support
+- Added marker based pagination for get users methods
+
 ## 2.41.0 [2019-10-24]
 - Added enum action option for completed in Box Task class.